WP-7: Transporting business logic to service.py
